refactor: :truck: move hooks para pasta especifica
